# Health check delay