--- a/src/server.rs
+++ b/src/server.rs
@@ -61,7 +61,7 @@ pub struct Server {
     application_name: String,
 
     // Last time that a successful server send or response happened
-    latest_successful_server_interaction_timestamp: SystemTime,
+    last_healthcheck: SystemTime,
 }
 
 impl Server {
@@ -320,7 +320,7 @@ impl Server {
                         connected_at: chrono::offset::Utc::now().naive_utc(),
                         stats: stats,
                         application_name: String::new(),
-                        latest_successful_server_interaction_timestamp: SystemTime::now(),
+                        last_healthcheck: SystemTime::now(),
                     };
 
                     server.set_name("pgcat").await?;
@@ -372,7 +372,8 @@ impl Server {
 
         match write_all_half(&mut self.write, messages).await {
             Ok(_) => {
-                self.latest_successful_server_interaction_timestamp = SystemTime::now();
+                // Successfully sent to server
+                self.last_healthcheck = SystemTime::now();
                 Ok(())
             }
             Err(err) => {
@@ -482,7 +483,8 @@ impl Server {
         // Clear the buffer for next query.
         self.buffer.clear();
 
-        self.latest_successful_server_interaction_timestamp = SystemTime::now();
+        // Successfully received data from server
+        self.last_healthcheck = SystemTime::now();
 
         // Pass the data back to the client.
         Ok(bytes)
@@ -576,8 +578,8 @@ impl Server {
     }
 
     // Get server's latest response timestamp
-    pub fn latest_successful_server_interaction_timestamp(&self) -> SystemTime {
-        self.latest_successful_server_interaction_timestamp
+    pub fn last_healthcheck(&self) -> SystemTime {
+        self.last_healthcheck
     }
 }