Currently we have introduced latency in transaction mode because we run health checks every time a connection is checked out from this pool.

This PR addresses this issue by only performing a health check after a certain amount of time has elapsed since the last time the server behaved normally. It also adds shard banning logic to the client side for handling errors communicating with the server connection served by the pool since they may not have passed health checks.