@zainkabani 2022-08-10
Performance improvements:

Before PR:
```
$ pgbench -t 25000 -c 16 -j 2 -S --protocol extended
pgbench (14.4, server 14.2 (Debian 14.2-1.pgdg110+1))
starting vacuum...end.
transaction type: <builtin: select only>
scaling factor: 1
query mode: extended
number of clients: 16
number of threads: 2
number of transactions per client: 25000
number of transactions actually processed: 400000/400000
latency average = 4.453 ms
initial connection time = 18.360 ms
tps = 3593.118431 (without initial connection time)
```

After PR:
```
$ pgbench -t 25000 -c 16 -j 2 -S --protocol extended
pgbench (14.4, server 14.2 (Debian 14.2-1.pgdg110+1))
starting vacuum...end.
transaction type: <builtin: select only>
scaling factor: 1
query mode: extended
number of clients: 16
number of threads: 2
number of transactions per client: 25000
number of transactions actually processed: 400000/400000
latency average = 2.499 ms
initial connection time = 10.946 ms
tps = 6403.132002 (without initial connection time)
```